% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_codes}
\alias{emoji_codes}
\title{Markdown emoji codes}
\format{A vector of length 1,435}
\source{
\href{https://github.com/ikatyang/emoji-cheat-sheet/blob/master/README.md}{emoji-cheat-sheet.md}
}
\usage{
emoji_codes
}
\description{
A names vector of emojis and their markdown codes.
}
\keyword{datasets}
