% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluedown.R
\docType{package}
\name{gluedown}
\alias{gluedown}
\alias{gluedown-package}
\title{gluedown: A package to format character vectors with markdown.}
\description{
The gluedown package provides functions to create both container blocks and
inlines from regular R character vectors. The package also helps uses these
new formatted vectors as actual text in a markdown document.
}
