% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_knitr}
\alias{has_knitr}
\alias{has_markdown}
\title{Check if the user has the 'knitr' package installed}
\usage{
has_knitr()

has_markdown()
}
\description{
Check if the user has the 'knitr' package installed
}
