% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-cat.R
\name{if_cat}
\alias{if_cat}
\title{Output if Desired}
\usage{
if_cat(x, cat)
}
\arguments{
\item{x}{The wrapped vector from a parent function.}

\item{cat}{The \code{cat} argument from a parent function.}
}
\value{
A single character vector formatted for printing.
}
\description{
Output if Desired
}
