% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-break.R
\name{md_break}
\alias{md_break}
\title{Markdown Thematic Break}
\usage{
md_break(char = c("*", "-", "_"), n = 3)
}
\arguments{
\item{char}{The character used. Either an asterisk, hyphen, or underscore.}

\item{n}{The number of times to repeat each character.}
}
\value{
A character vector of a repeated horizontal rule character.
}
\description{
Markdown Thematic Break
}
\details{
From the \href{https://github.github.com/gfm/}{GFM spec}: A line
consisting of 0-3 spaces of indentation, followed by a sequence of three or
more matching \code{-}, \code{_}, or \code{*} characters, each followed optionally by any
number of spaces or tabs, forms a thematic break.
}
\examples{
md_break("_")
md_break(n = 10)
}
