% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-define.R
\name{md_define}
\alias{md_define}
\title{Markdown definitions}
\usage{
md_define(term, def)
}
\arguments{
\item{term}{A character vector of terms to define.}

\item{def}{A character vector of definitions to be used for each \code{term.}}
}
\value{
A character vector with elements preceded by an asterisk symbol.
}
\description{
Form a single-level unordered list from a vector.
}
\details{
Definitions are not supported in GFM, but \emph{are} supported by the
\href{https://pandoc.org/MANUAL.html#pandocs-markdown}{pandoc engine} used by R
Markdown. Per that specification: Each term must fit on one line, which may
optionally be followed by a blank line, and must be followed by one or more
definitions. A definition begins with a colon or tilde, which may be indented
one or two spaces.
}
\examples{
md_define("Democracy", "Government by the people")
md_define(state.abb[1:5], state.name[1:5])
}
