% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_emoji.R
\name{md_emoji}
\alias{md_emoji}
\title{Markdown emoji (unsupported extension)}
\usage{
md_emoji(x)
}
\arguments{
\item{x}{A character vector containing emoji codes in the format of \code{:code:}.}
}
\value{
A \code{glue} vector of emoji?
}
\description{
Replace an emoji text code with it's UTF-8 counterpart.
}
\examples{
md_emoji(":sunglasses:")
md_emoji(c(":man_technologist:", ":arrow_down:"))
}
