% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasize.R
\name{md_emphasize}
\alias{md_emphasize}
\title{Markdown Selective Emphasize}
\usage{
md_emphasize(x, f, n)
}
\arguments{
\item{x}{A vectror of sentence character strings.}

\item{f}{An emphasizing function (e.g., \code{\link[=md_bold]{md_bold()}}).}

\item{n}{The numeric \emph{position} of the word in each sentence to emphasize.}
}
\value{
A character vector with elements wrapped in two tilde symbols.
}
\description{
Markdown Selective Emphasize
}
\examples{
md_emphasize(x = "Live free or die", f = md_bold, n = 4)
}
