% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-escape.R
\name{md_escape}
\alias{md_escape}
\title{Backslash escape all punctuation}
\usage{
md_escape(x)
}
\arguments{
\item{x}{A character vector of strings containing punctuation that might
accidentally be considered markdown syntax.}
}
\value{
A character string with all \verb{[:punct:]} properly escaped with
prepended backslashes.
}
\description{
Take a character vector containing punctuation and return a glue vector with
every punctuation mark prepended with double escape backslashes.
}
\details{
When trying to format text containing markdown syntax characters, it's
necessary to "escape" those characters so that they are ignored by
formatting.

Any ASCII punctuation character may be backslash-escaped... Escaped
characters are treated as regular characters and do not have their usual
Markdown meanings.
}
\examples{
md_escape("# six seasons and a movie")
}
\seealso{
Other inline functions:
\code{\link[=md_autolink]{md_autolink()}},
\code{\link[=md_bold]{md_bold()}},
\code{\link[=md_code]{md_code()}},
\code{\link[=md_convert]{md_convert()}},
\code{\link[=md_disallow]{md_disallow()}},
\code{\link[=md_hardline]{md_hardline()}},
\code{\link[=md_image]{md_image()}},
\code{\link[=md_issue]{md_issue()}},
\code{\link[=md_italic]{md_italic()}},
\code{\link[=md_link]{md_link()}},
\code{\link[=md_softline]{md_softline()}},
\code{\link[=md_strike]{md_strike()}},
\code{\link[=md_text]{md_text()}}
}
\concept{inline functions}
