% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-wrap.R
\name{md_header}
\alias{md_header}
\title{Markdown ATX Headings}
\usage{
md_header(x, level = 1)
}
\arguments{
\item{x}{The text of the headers.}

\item{level}{The number of header hash characters. The heading level is equal
to the number of # characters in the opening sequence. In rendered markdown
higher levels are typically displayed larger.}
}
\value{
A character vector with elements preceded by hash symbols.
}
\description{
Markdown ATX Headings
}
\examples{
md_header("Overview")
}
