% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-wrap.R
\name{md_inline}
\alias{md_inline}
\title{Markdown Inline Code}
\usage{
md_inline(x)
}
\arguments{
\item{x}{The text to be formatted as inline code.}
}
\value{
A character vector with elements wrapped in backtick symbols.
}
\description{
Markdown Inline Code
}
\examples{
md_inline("ex_var")
md_inline(state.name)
}
