% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-math.R
\name{md_math}
\alias{md_math}
\title{Markdown Math Equations}
\usage{
md_math(x)
}
\arguments{
\item{x}{The text to formatted as math.}
}
\value{
A character vector with elements wrapped in a single dollar sign.
}
\description{
Markdown Math Equations
}
\examples{
md_math("1 + 1")
}
