% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-text.R
\name{md_text}
\alias{md_text}
\title{Markdown textual content}
\usage{
md_text(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A \code{glue} vector.
}
\description{
Simple wrapper around \code{\link[glue:as_glue]{glue::as_glue()}}. Take a character vector and return a
glue vector.
}
\details{
Any characters not given an interpretation by the [other] rules will be
parsed as plain textual content.
}
\examples{
md_text("foo")
}
\seealso{
Other inline functions:
\code{\link[=md_autolink]{md_autolink()}},
\code{\link[=md_bold]{md_bold()}},
\code{\link[=md_code]{md_code()}},
\code{\link[=md_convert]{md_convert()}},
\code{\link[=md_disallow]{md_disallow()}},
\code{\link[=md_escape]{md_escape()}},
\code{\link[=md_hardline]{md_hardline()}},
\code{\link[=md_image]{md_image()}},
\code{\link[=md_issue]{md_issue()}},
\code{\link[=md_italic]{md_italic()}},
\code{\link[=md_link]{md_link()}},
\code{\link[=md_softline]{md_softline()}},
\code{\link[=md_strike]{md_strike()}}
}
\concept{inline functions}
