% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasize.R
\name{emphasize_at}
\alias{emphasize_at}
\title{Selective Markdown Emphasize}
\usage{
emphasize_at(x, f, n)
}
\arguments{
\item{x}{A vector of sentence character strings.}

\item{f}{An emphasizing function (e.g., \code{\link[=md_bold]{md_bold()}}).}

\item{n}{The numeric \emph{position} of the word in each sentence to emphasize.}
}
\value{
A character vector with elements wrapped in two tilde symbols.
}
\description{
Selective Markdown Emphasize
}
\examples{
emphasize_at(x = "Live free or die", n = 4, f = md_bold)
}
