% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasize-at.R
\name{emphasize_at}
\alias{emphasize_at}
\title{Targeted markdown formatting}
\usage{
emphasize_at(x, f, n)
}
\arguments{
\item{x}{A character vector of strings containing words separated by spaces.}

\item{f}{An emphasizing inline function (e.g., \code{\link[=md_bold]{md_bold()}}).}

\item{n}{A numeric vector describing the positions of the words in \code{x} to be
emphasized with \code{f}.}
}
\value{
A \code{glue} vector of length equal to \code{x} with the \code{n}th word
formatted using \code{f}.
}
\description{
Split a sentence by whitespace and wrap a specific word(s) in a
markdown inline function.
}
\examples{
emphasize_at(x = "Live free or die", n = 4, f = md_bold)
}
