% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-lines.R
\name{md_blank}
\alias{md_blank}
\title{Markdown blank line}
\usage{
md_blank()
}
\value{
A \code{glue} vector of length one containing two newline characters.
}
\description{
Create a blank line between other markdown block-level elements.
}
\details{
Blank lines between block-level elements are ignored, except for the role
they play in determining whether a list is tight or loose.

Blank lines at the beginning and end of the document are also ignored.
}
\examples{
md_blank
}
\seealso{
Other leaf block functions: 
\code{\link{md_chunk}()},
\code{\link{md_fence}()},
\code{\link{md_heading}()},
\code{\link{md_indent}()},
\code{\link{md_label}()},
\code{\link{md_paragraph}()},
\code{\link{md_reference}()},
\code{\link{md_rule}()},
\code{\link{md_setext}()},
\code{\link{md_table}()}
}
\concept{leaf block functions}
