% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasize.R
\name{md_bold}
\alias{md_bold}
\title{Markdown Bold Emphasis}
\usage{
md_bold(x)
}
\arguments{
\item{x}{The text to be emphasized in bold.}
}
\value{
A character vector with elements wrapped in two asterisk symbols.
}
\description{
Markdown Bold Emphasis
}
\examples{
md_bold("Example")
md_bold(state.name)
}
\seealso{
Other inline functions: \code{\link{md_code}},
  \code{\link{md_escape}}, \code{\link{md_image}},
  \code{\link{md_issue}}, \code{\link{md_italic}},
  \code{\link{md_link}}, \code{\link{md_strike}}
}
\concept{inline functions}
