% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-wrap.R
\name{md_bold}
\alias{md_bold}
\title{Markdown Bold Emphasis}
\usage{
md_bold(x)
}
\arguments{
\item{x}{The text to be emphasized in bold.}
}
\value{
A character vector with elements wrapped in two asterisk symbols.
}
\description{
Markdown Bold Emphasis
}
\examples{
md_bold("Example")
md_bold(state.name)
}
