% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-list.R
\name{md_bullet}
\alias{md_bullet}
\title{Markdown Bullet List}
\usage{
md_bullet(x, marker = c("*", "-", "+"))
}
\arguments{
\item{x}{The vector of bullet point list items.}

\item{marker}{The bullet list marker to use.}
}
\value{
A character vector with elements preceded by an asterisk symbol.
}
\description{
Form a single-level unordered list from a vector.
}
\details{
From the \href{https://github.github.com/gfm/#bullet-list-marker}{GFM spec}:
A bullet list marker is a \code{-}, \code{+}, or \code{*} character.
}
\examples{
md_bullet(state.name[1:5])
md_bullet(state.name[1:5])
}
