% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-format.R
\name{md_bullet}
\alias{md_bullet}
\title{Markdown Bullet List}
\usage{
md_bullet(x, cat = FALSE)
}
\arguments{
\item{x}{The vector of bullet point list items.}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to `FALSE`.}
}
\value{
A character vector with elements preceded by an asterisk symbol.
}
\description{
Markdown Bullet List
}
\examples{
md_bullet(state.name)
md_bullet(state.name, cat = TRUE)
}
