% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-cat.R
\name{md_cat}
\alias{md_cat}
\title{Output Separated Markdown}
\usage{
md_cat(x)
}
\arguments{
\item{x}{The markdown text to concatenate and separate with newlines.}
}
\value{
A single character vector formatted for printing.
}
\description{
Output Separated Markdown
}
\examples{
md_cat(md_list(c("one", "two")))
}
