% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-fence.R
\name{md_chunk}
\alias{md_chunk}
\title{Markdown Code Block}
\usage{
md_chunk(lines, lang = "r", type = c("tick", "tilde", "indent"),
  style = TRUE)
}
\arguments{
\item{lines}{A character vector of lines to
be wrapped in fences, possibly created by \code{\link[readr:read_lines]{readr::read_lines()}} or some
equivalent.}

\item{lang}{The language of the code within the fence. Aside from clarity,
this allows many markdown engines to adjust the syntax highlighting.
Defaults to "r". This option is not used with indented blocks.}

\item{type}{The type of code block to be created. Defaults to tipple
backticks code fences, but you can also use 4-spaces (useful for legacy
support on Reddit).}

\item{style}{logical; If the \code{styler} package is installed, should
\code{\link[styler:style_text]{styler::style_text()}} be called on \code{lines}?}
}
\value{
A character vector wrapped on either side by code fences.
}
\description{
Wrap a character vector (possibly lines of a file) in code
fences to be rendered as a code block.
}
\details{
From the \href{https://github.github.com/gfm/#code-fence}{GFM spec}: A
code fence is a sequence of at least three consecutive backtick characters…
A fenced code block begins with a code fence, indented no more than three
spaces… The line with the opening code fence may optionally contain some
text following the code fence… The content of the code block consists of
all subsequent lines, until a closing code fence… The content of a code
fence is treated as literal text, not parsed as inlines. The first word of
the info string is typically used to specify the language of the code
sample
}
\examples{
md_chunk(c("library(ggplot2)", "ggplot(mpg)+", "geom_point(aes(displ, hwy))"))
md_chunk(c("library(dplyr)", "starwars \%>\%", "filter(species == 'Droid')"), type = "indent")
}
