% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasis.R
\name{md_code}
\alias{md_code}
\title{Markdown Inline Code}
\usage{
md_code(x)
}
\arguments{
\item{x}{The text to be formatted as fixed-width inline code.}
}
\value{
A character vector with elements wrapped in backtick symbols.
}
\description{
Markdown Inline Code
}
\examples{
md_code("ex_var")
md_code(state.name)
}
