% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasize.R
\name{md_code}
\alias{md_code}
\title{Markdown Inline Code}
\usage{
md_code(x)
}
\arguments{
\item{x}{The text to be formatted as fixed-width inline code.}
}
\value{
A character vector with elements wrapped in backtick symbols.
}
\description{
Markdown Inline Code
}
\examples{
md_code("ex_var")
md_code(state.name)
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_escape}}, \code{\link{md_image}},
  \code{\link{md_issue}}, \code{\link{md_italic}},
  \code{\link{md_link}}, \code{\link{md_strike}}
}
\concept{inline functions}
