% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-convert.R
\name{md_convert}
\alias{md_convert}
\title{Convert Markdown to HTML}
\usage{
md_convert(md, frag = TRUE)
}
\arguments{
\item{md}{The markdown text to be converted.}

\item{frag}{logical; Whether only a single HTML fragment should be returned.}
}
\value{
The HTML fragment equivalent.
}
\description{
Wraps around \code{\link[markdown:markdownToHTML]{markdown::markdownToHTML()}} to convert markdown
text into HTML \emph{fragments}.
}
\examples{
md_convert(md = md_bold("Test"))
}
