% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-convert.R
\name{md_convert}
\alias{md_convert}
\title{Convert markdown to HTML}
\usage{
md_convert(x, frag = TRUE)
}
\arguments{
\item{x}{A character vector of \emph{markdown} text to be converted.}

\item{frag}{logical; Whether only a single HTML fragment should be returned.
\code{TRUE} by default.}
}
\value{
A \code{glue} vector of length 1 containing HTML tags.
}
\description{
Take a character vector of valid markdown text and pass it to
\code{\link[markdown:markdownToHTML]{markdown::markdownToHTML()}} to create a glue vector of HTML fragments.
Primarily used to test that \code{md_*()} functions create vectors that meet the
GFM spec and can be rendered as HTML.
}
\examples{
md_convert(x = md_bold("test"))
}
