% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-define.R
\name{md_define}
\alias{md_define}
\title{Markdown Definition}
\usage{
md_define(term, def, cat = TRUE)
}
\arguments{
\item{term}{The chracter vector of terms to define.}

\item{def}{The character vector of definitions to be used for each term.}

\item{cat}{logical; Should the list be concatenated and printed, with each
definition entry separated by a new line? Defaults to \code{TRUE}.}
}
\value{
A character vector with elements preceded by an asterisk symbol.
}
\description{
Form a single-level unordered list from a vector.
}
\details{
Defintions are not supported in GFM, but \emph{are} supported by the
\href{https://pandoc.org/MANUAL.html#pandocs-markdown}{pandoc engine} used by R
Markdown. Per that specification: Each term must fit on one line, which
may optionally be followed by a blank line, and must be followed by one or
more definitions. A definition begins with a colon or tilde, which may be
indented one or two spaces.
}
\examples{
md_define("Democracy", "Government by the people")
}
