% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-escape.R
\name{md_escape}
\alias{md_escape}
\title{Markdown Escape Punctuation}
\usage{
md_escape(x)
}
\arguments{
\item{x}{The text with punctuation to escape.}
}
\value{
A character string with all \code{[:punct:]} properly escaped with
prepended backslashes.
}
\description{
When trying to format text containing markdown syntax
characters, it's necessary to "escape" those characters so that they are
ignored by formatting.
}
\details{
From the \href{https://github.github.com/gfm/#backslash-escapes}{GFM spec}:
Any ASCII punctuation character may be backslash-escaped... Any ASCII
punctuation character may be backslash-escaped... Escaped characters are
treated as regular characters and do not have their usual Markdown meanings.
}
\examples{
md_bold(md_escape("Use an * symbol in globs!"))
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_hardline}},
  \code{\link{md_image}}, \code{\link{md_issue}},
  \code{\link{md_italic}}, \code{\link{md_link}},
  \code{\link{md_strike}}
}
\concept{inline functions}
