% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-wrap.R
\name{md_fence}
\alias{md_fence}
\title{Markdown Code Fence}
\usage{
md_fence(x, lang = "R", cat = TRUE)
}
\arguments{
\item{x}{Either (1) a path of _one_ existing file to be read as a character
vector or (2) a character vector of lines to be wrapped in fences.}

\item{lang}{The language of the code within the fence. Aside from clarity,
this allows many markdown engines to adjust the syntax highlighing.
Defaults to "R".}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to `TRUE`.}
}
\value{
A character vector wrapped on either side by code fences.
}
\description{
Markdown Code Fence
}
\examples{
md_fence(gluedown:::plot)
}
