% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-fence.R
\name{md_fence}
\alias{md_fence}
\title{Markdown Code Fence}
\usage{
md_fence(x, lang = "r", cat = TRUE)
}
\arguments{
\item{x}{A character vector of lines (with spaces for indentation) to be wrapped in fences.}

\item{lang}{The language of the code within the fence. Aside from clarity,
this allows many markdown engines to adjust the syntax highlighing.
Defaults to "R".}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to \code{TRUE}.}
}
\value{
A character vector wrapped on either side by code fences.
}
\description{
Wrap a character vector (possibly lines of a file) in code fences to be rendered as
a code block.
}
\details{
From the \href{https://github.github.com/gfm/#code-fence}{GFM spec}: A code fence is a
sequence of at least three consecutive backtick characters… A fenced code block begins with a
code fence, indented no more than three spaces… The line with the opening code fence may
optionally contain some text following the code fence… The content of the code block consists
of all subsequent lines, until a closing code fence… The content of a code fence is treated as
literal text, not parsed as inlines. The first word of the info string is typically used to
specify the language of the code sample
}
\examples{
md_fence(c("library(ggplot2)", "ggplot(data = mpg) +", "  geom_point((x = displ, y = hwy))"))
}
