% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-chunk.R
\name{md_fence}
\alias{md_fence}
\title{Markdown fenced code block (4.5)}
\usage{
md_fence(x, char = c("`", "~"), lang = "r")
}
\arguments{
\item{lang}{The language of the code within the fence. Aside from clarity,
this allows many markdown engines to adjust the syntax highlighting.
Defaults to "r". This option is not used with indented blocks.}

\item{lines}{A character vector of lines to be wrapped concatenated into a
single block, possibly created by \code{\link[readr:read_lines]{readr::read_lines()}} or some equivalent.}

\item{type}{The type of code block to be created. Defaults to tipple
backticks code fences, but you can also use 4-spaces (useful for legacy
support on Reddit).}
}
\value{
A character vector wrapped on either side by code fences.
}
\description{
Turn a character vector of lines into a single code block with
lines bookended with a code fence of backticks or tildes. This markdown
leaf block can be rendered as HTML \code{<code>} tags inside \code{<pre>} tags.
}
\details{
From the \href{https://github.github.com/gfm/#code-fence}{GFM spec}: A
code fence is a sequence of at least three consecutive backtick characters
(\code{`}) or tildes (\code{~}). (Tildes and backticks cannot be mixed.) A fenced
code block begins with a code fence, indented no more than three spaces.

The line with the opening code fence may optionally contain some text
following the code fence; this is trimmed of leading and trailing
whitespace and called the info string (see \code{lang})...

The content of the code block consists of all subsequent lines, until a
closing code fence of the same type as the code block began with (backticks
or tildes), and with at least as many backticks or tildes as the opening
code fence...

The content of a code fence is treated as literal text, not parsed as
inlines. The first word of the info string is typically used to specify the
language of the code sample, and rendered in the class attribute of the
code tag.
}
\examples{
md_fence(deparse(sd))
md_fence(c("library(dplyr)", "starwars \%>\%", "  filter(species == 'Droid')"))
}
\seealso{
Other leaf block functions: \code{\link{md_chunk}},
  \code{\link{md_heading}}, \code{\link{md_indent}},
  \code{\link{md_reference}}, \code{\link{md_rule}},
  \code{\link{md_setext}}, \code{\link{md_table}}
}
\concept{leaf block functions}
