% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-hardline.R
\name{md_hardline}
\alias{md_hardline}
\title{Markdown hard line breaks (6.12)}
\usage{
md_hardline(x)
}
\arguments{
\item{x}{A character vector of bullet point list items.}
}
\value{
A \code{glue} vector of length 1 with elements of \code{x} separated by two
spaces and a newline.
}
\description{
Separate strings with a hard line (non-paragraph) break. This inline hard
line break can be rendered as \code{<br />} HTML tags.
}
\details{
A line break (not in a code span or HTML tag) that is preceded by two or more
spaces and does not occur at the end of a block is parsed as a hard line
break (rendered in HTML as a \code{<br />} tag)
}
\examples{
md_hardline(md_bold(c("One", "Two")))
md_hardline(state.name[1:10])
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_escape}},
  \code{\link{md_image}}, \code{\link{md_issue}},
  \code{\link{md_italic}}, \code{\link{md_link}},
  \code{\link{md_strike}}
}
\concept{inline functions}
