% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-heading.R
\name{md_heading}
\alias{md_heading}
\title{Markdown ATX Headings}
\usage{
md_heading(x, level = 1, cat = FALSE)
}
\arguments{
\item{x}{The text of the headings.}

\item{level}{The number of header hash characters. The heading level is equal
to the number of # characters in the opening sequence. In rendered markdown
higher levels are typically displayed larger.}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to \code{FALSE}.}
}
\value{
A character vector with elements preceded by hash symbols.
}
\description{
Form a vector of markdown headings.
}
\details{
From the \href{https://github.github.com/gfm/#atx-heading}{GFM spec}: An
ATX heading consists of a string of characters, parsed as inline content,
between an opening sequence of 1–6 unescaped \code{#} characters… The opening
sequence of \code{#} characters must be followed by a space or by the end of
line… The heading level is equal to the number of \code{#} characters in the
opening sequence.
}
\examples{
md_heading("Overview")
md_heading(c("Installation", "Development version"), level = 1:2)
}
