% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-link.R
\name{md_image}
\alias{md_image}
\title{Markdown Image}
\usage{
md_image(url, alt = "")
}
\arguments{
\item{url}{The URL of an image to be displayed or referenced.}

\item{alt}{The optional
\href{https://en.wikipedia.org/wiki/Alt_attribute}{alternative text} that can
be used to refer to an image.}
}
\value{
A character vector of collapsed alternative text and associated URLs.
}
\description{
Markdown Image
}
\examples{
md_image("http://hexb.in/hexagons/eff.png")
md_image("http://hexb.in/hexagons/eff.png", "EFF Hex Sticker")
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_escape}},
  \code{\link{md_issue}}, \code{\link{md_italic}},
  \code{\link{md_link}}, \code{\link{md_strike}}
}
\concept{inline functions}
