% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-format.R
\name{md_issue}
\alias{md_issue}
\title{Markdown Issue}
\usage{
md_issue(repo, num)
}
\arguments{
\item{repo}{A character vector in the format `"user/rep"`.}

\item{num}{The issue or pull number _without_ hash symbol.}
}
\value{
A character vector which GitHub can automatically hyperlink.
}
\description{
A markdown feature unique to GitHub. Allows users to hyperlink
a specific issue or pull request using a username, repository name, and
issue number.
}
\examples{
md_issue("kiernann/gluedown", 1)
}
