% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-issue.R
\name{md_issue}
\alias{md_issue}
\title{Markdown Issue}
\usage{
md_issue(repo, num)
}
\arguments{
\item{repo}{A character vector in the format \code{"user/rep"}.}

\item{num}{The issue or pull number \emph{without} hash symbol.}
}
\value{
A character vector which GitHub can automatically hyperlink.
}
\description{
A markdown feature unique to GitHub. Allows users to hyperlink
a specific issue or pull request using a username, repository name, and
issue number.
}
\examples{
md_issue("kiernann/gluedown", 1)
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_escape}},
  \code{\link{md_hardline}}, \code{\link{md_image}},
  \code{\link{md_italic}}, \code{\link{md_link}},
  \code{\link{md_strike}}

Other markdown extensions: \code{\link{md_table}}
}
\concept{inline functions}
\concept{markdown extensions}
