% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-inlines.R
\name{md_italic}
\alias{md_italic}
\title{Markdown italic emphasis (6.4)}
\usage{
md_italic(x)
}
\arguments{
\item{x}{The text to be emphasized in italics.}
}
\value{
A \code{glue} vector of length equal to \code{x}.
}
\description{
Turn a character vector into a vector of valid italic emphasis
markdown inline text. This inline text can then be rendered as the \code{<em>}
HTML tag. Right now, this package uses sinlge \code{_} for strong emphasis,
although single \code{*} meets the GFM spec.
}
\details{
A single \code{*} or \code{_} can open or close emphasis... Emphasis begins
with a delimiter that can open emphasis and ends with a delimiter that can
close emphasis, and that uses the same character (\code{_} or \code{*}) as the
opening delimiter.
}
\examples{
md_italic("Example")
md_italic(state.name)
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_escape}},
  \code{\link{md_hardline}}, \code{\link{md_image}},
  \code{\link{md_issue}}, \code{\link{md_link}},
  \code{\link{md_strike}}
}
\concept{inline functions}
