% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasize.R
\name{md_italic}
\alias{md_italic}
\title{Markdown Italic Emphasis}
\usage{
md_italic(x)
}
\arguments{
\item{x}{The text to be emphasized in italics.}
}
\value{
A character vector with elements wrapped in underscore symbols.
}
\description{
Markdown Italic Emphasis
}
\examples{
md_italic("Example")
md_italic(state.name)
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_escape}},
  \code{\link{md_image}}, \code{\link{md_issue}},
  \code{\link{md_link}}, \code{\link{md_strike}}
}
\concept{inline functions}
