% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-link.R
\name{md_label}
\alias{md_label}
\title{Markdown link label}
\usage{
md_label(text, label, ..., .name = FALSE)
}
\arguments{
\item{label}{A link label that is referenced elsewhere in the document.}

\item{...}{A sequence of \code{text = "/url"} named vector pairs. If any such
pairs are provided, \code{.name} will be considered \code{TRUE}.}

\item{.name}{logical; if \code{TRUE}, the pairs in \code{...} will be used instead of
any values supplied to \code{x} and \code{url}.}

\item{url}{The URL to hyperlink the referenced text with.}

\item{title}{An \emph{optional} link title; defaults to \code{NULL}.}
}
\value{
A single \code{glue} vector of length equal to that of \code{label} and \code{url},
with elements the concatenated arguments.
}
\description{
Create the link labels that can latter be refered to with a link reference
from \code{\link[=md_reference]{md_reference()}}.
}
\details{
A link label begins with a left bracket (\code{[}) and ends with the first right
bracket (\verb{]}) that is not backslash-escaped. Between these brackets there
must be at least one non-whitespace character.

[\verb{) and ends with the first right bracket (}]: R:\%60)\%20and\%20ends\%20with\%20the\%20first\%20right\%0Abracket\%20(\%60
}
\examples{
md_label(CRAN = "The CRAN website")
md_label(text = c("one", "two"), label = 1:2)
}
\seealso{
Other leaf block functions: 
\code{\link{md_blank}()},
\code{\link{md_chunk}()},
\code{\link{md_fence}()},
\code{\link{md_heading}()},
\code{\link{md_indent}()},
\code{\link{md_paragraph}()},
\code{\link{md_reference}()},
\code{\link{md_rule}()},
\code{\link{md_setext}()},
\code{\link{md_table}()}
}
\concept{leaf block functions}
