% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-wrap.R
\name{md_link}
\alias{md_link}
\title{Markdown Hyperlink}
\usage{
md_link(text, url)
}
\arguments{
\item{text}{The chracter vector to be displayed as hyperlinked text.}

\item{url}{The URL to lead to.}
}
\value{
A character vector of collapsed display text and associated URLs.
}
\description{
Markdown Hyperlink
}
\examples{
md_link("tidyverse", "https://www.tidyverse.org/")
md_link(c("R Project", "CRAN"), c("https://www.r-project.org/", "https://cran.r-project.org/"))
}
