% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-format.R
\name{md_list}
\alias{md_list}
\title{Markdown Numered List}
\usage{
md_list(x, seq = TRUE, pad = TRUE, cat = FALSE)
}
\arguments{
\item{x}{The vector of numbered list items.}

\item{seq}{logical; Should sequential numbers be used? Defaults to `TRUE`. If
`FALSE`, each element will be preceded by the number one; many markdown
engines will automatically render repeated ones as a sequential list.}

\item{pad}{logical; If sequential numbers are used, should they be padded
with zeroes on the left to match the width of the greatest number? Defaults
to `TRUE`.}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to `FALSE`.}
}
\value{
A character vector with elements preceded by a number.
}
\description{
Markdown Numered List
}
\examples{
md_list(state.name)
md_list(state.name, cat = TRUE)
}
