% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-list.R
\name{md_list}
\alias{md_list}
\title{Markdown Numered List}
\usage{
md_list(x, seq = TRUE, pad = TRUE, punct = c(".", ")"),
  cat = FALSE)
}
\arguments{
\item{x}{The vector of numbered list items.}

\item{seq}{logical; Should sequential numbers be used? Defaults to \code{TRUE}. If
\code{FALSE}, each element will be preceded by the number one; many markdown
engines will automatically render repeated ones as a sequential list.}

\item{pad}{logical; If sequential numbers are used, should they be padded
with zeroes on the left to match the width of the greatest number? Defaults
to \code{TRUE}.}

\item{punct}{The puctuation mark following each number; either \code{.} or \code{)}.}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to \code{FALSE}.}
}
\value{
A character vector with elements preceded by a number.
}
\description{
Form a single-level ordered list from a vector.
}
\details{
From the \href{https://github.github.com/gfm/#ordered-list-marker}{GFM spec}: An ordered
list marker is a sequence of 1–9 arabic digits (\code{0-9}), followed by either a \code{.} character or
a \code{)} character.
}
\examples{
md_list(state.name[1:5])
md_list(state.name[1:5], cat = TRUE)
}
