% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-format.R
\name{md_quote}
\alias{md_quote}
\title{Markdown Block Quotes}
\usage{
md_quote(x)
}
\arguments{
\item{x}{The character vector of quotes.}
}
\value{
A character vector with elements preceded by an greater-than symbol.
}
\description{
Markdown Block Quotes
}
\examples{
md_quote(gluedown:::preamble)
}
