% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-quote.R
\name{md_quote}
\alias{md_quote}
\title{Markdown Block Quotes}
\usage{
md_quote(x)
}
\arguments{
\item{x}{The character vector of quotes.}
}
\value{
A character vector with a greater-than symbol (\code{>}) prepended to each
element.
}
\description{
Form block quotes from a vector of character strings.
}
\details{
From the \href{https://github.github.com/gfm/#block-quote-marker}{GFM spec}:
A block quote marker consists of 0-3 spaces of initial indent, plus (a) the
character \code{>} together with a following space, or (b) a single character
\code{>} not followed by a space.
}
\examples{
md_quote("Ask not what your country can do for you, but what you can do for your country.")
}
