% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-format.R
\name{md_rule}
\alias{md_rule}
\title{Markdown Horizontal Rule}
\usage{
md_rule(char = c("*", "-", "_"), n = 3)
}
\arguments{
\item{char}{The character used. Either an asterisk, hyphen, or underscore.}

\item{n}{The number of times to repeat each character.}
}
\value{
A character vector of a repeated horizontal rule character.
}
\description{
Markdown Horizontal Rule
}
\examples{
md_rule()
md_rule("_", n = 10)
}
