% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-break.R
\name{md_rule}
\alias{md_rule}
\title{Markdown horizontal rule (4.1)}
\usage{
md_rule(char = c("*", "-", "_"), n = 3)
}
\arguments{
\item{char}{The type of rule; either: \code{-}, \code{_}, or \code{*}. Defaults to \code{*}.}

\item{n}{The width of the rule; an integer indicating number of times to
repeat each character. Defaults to the minimum of 3.}
}
\value{
A repeated-character \code{glue} vector with length 1.
}
\description{
Print the markdown characters used to represent a \emph{thematic
break}. This markdown leaf block can be rendered as the \code{<hr>} HTML tag.
}
\details{
A line consisting of 0-3 spaces of indentation, followed by a
sequence of three or more matching \code{-}, \code{_}, or \code{*} characters, each
followed optionally by any number of spaces or tabs, forms a thematic
break.
}
\examples{
md_rule()
md_rule("_")
md_rule(n = 10)
}
\seealso{
Other leaf block functions: \code{\link{md_chunk}},
  \code{\link{md_fence}}, \code{\link{md_heading}},
  \code{\link{md_indent}}, \code{\link{md_reference}},
  \code{\link{md_setext}}, \code{\link{md_table}}
}
\concept{leaf block functions}
