% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-heading.R
\name{md_setext}
\alias{md_setext}
\title{Markdown Setext headings (4.3)}
\usage{
md_setext(x, level = 1)
}
\arguments{
\item{x}{A character vector of heading text.}

\item{level}{A numeric vector of use to determine the number of heading hash
characters to preceed each element of \code{x}. The heading level is equal to
the number of \code{#} characters in the opening sequence.}
}
\value{
A \code{glue} vector of a length equal to \code{x}.
}
\description{
Turn a character vector into a vector of valid markdown Setext
headings. These markdown leaf blocks can be rendered as the \code{<h1>} and
\code{<h2>} tags \emph{only}.
}
\details{
From the \href{https://github.github.com/gfm/#setext-headings}{GFM spec}:
A setext heading consists of one or more lines of text, each containing at
least one \href{https://github.github.com/gfm/#non-whitespace-character}{non-whitespace character},
with no more than 3 spaces indentation,
followed by a \href{https://github.github.com/gfm/#setext-heading-underline}{setext heading underline}...
A setext heading underline is a sequence of \code{=} characters or a sequence of
\code{-} characters, with no more than 3 spaces indentation and any number of
trailing spaces... The heading is a level 1 heading if \code{=} characters are
used in the setext heading underline, and a level 2 heading if \code{-}
characters are used.
}
\examples{
md_setext("Overview")
md_setext("This is a setext\\nheading", level = 2)
md_setext(c("one", "two", "three", "four"), level = c(1, 2))
}
\seealso{
Other leaf block functions: \code{\link{md_chunk}},
  \code{\link{md_fence}}, \code{\link{md_heading}},
  \code{\link{md_indent}}, \code{\link{md_reference}},
  \code{\link{md_rule}}, \code{\link{md_table}}
}
\concept{leaf block functions}
