% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-heading.R
\name{md_setext}
\alias{md_setext}
\title{Markdown Setext Headings}
\usage{
md_setext(x, level = 1)
}
\arguments{
\item{x}{The text of the headings.}

\item{level}{The heading level created; level 1 uses \code{=} and level 2 or
greater uses \code{-}.}
}
\value{
A heading with a setext underline the width of the longest line in \code{x}.
}
\description{
Form a single level 1 or 2 markdown setext headings.
}
\details{
From the \href{https://github.github.com/gfm/#setext-headings}{GFM spec}:
A setext heading consists of one or more lines of text, each containing at
least one \href{https://github.github.com/gfm/#non-whitespace-character}{non-whitespace character},
with no more than 3 spaces indentation,
followed by a \href{https://github.github.com/gfm/#setext-heading-underline}{setext heading underline}...
A setext heading underline is a sequence of \code{=} characters or a sequence of
\code{-} characters, with no more than 3 spaces indentation and any number of
trailing spaces... The heading is a level 1 heading if \code{=} characters are
used in the setext heading underline, and a level 2 heading if \code{-}
characters are used.
}
\examples{
md_setext("Overview")
md_setext("This is a setext\\nHeading", level = 2)
md_setext(c("one", "two", "three"), c(1, 2, 1))
}
\seealso{
Other leaf block functions: \code{\link{md_heading}}
}
\concept{leaf block functions}
