% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-format.R
\name{md_strike}
\alias{md_strike}
\title{Markdown Strikethough Emphasis}
\usage{
md_strike(x)
}
\arguments{
\item{x}{The text to be striked though.}
}
\value{
A character vector with elements wrapped in two tilde symbols.
}
\description{
Markdown Strikethough Emphasis
}
\examples{
md_strike("Example")
md_strike(state.name)
}
