% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-emphasize.R
\name{md_strike}
\alias{md_strike}
\title{Markdown Strikethrough Emphasis}
\usage{
md_strike(x)
}
\arguments{
\item{x}{The text to be striked though.}
}
\value{
A character vector with elements wrapped in two tilde symbols.
}
\description{
Markdown Strikethrough Emphasis
}
\examples{
md_strike("Example")
md_strike(state.name)
}
\seealso{
Other inline functions: \code{\link{md_bold}},
  \code{\link{md_code}}, \code{\link{md_escape}},
  \code{\link{md_image}}, \code{\link{md_issue}},
  \code{\link{md_italic}}, \code{\link{md_link}}
}
\concept{inline functions}
