% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-table.R
\name{md_table}
\alias{md_table}
\title{Markdown Tables}
\usage{
md_table(df, ...)
}
\arguments{
\item{df}{A data frame to be formatted as table.}
}
\value{
A character vector of markdown table source code.
}
\description{
Wraps around the [knitr::kable()] function.
}
\examples{
md_table(ggplot2::presidential)
}
