% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-wrap.R
\name{md_task}
\alias{md_task}
\title{Markdown Task List}
\usage{
md_task(x, check = NULL, cat = FALSE)
}
\arguments{
\item{x}{The vector of task list items.}

\item{check}{A vector of list elements which should be checked off.}

\item{cat}{logical; Should the list be concatenated and printed, with each
bullet element separated by a new line? Defaults to `FALSE`.}
}
\value{
A character vector with elements preceded check box brackets.
}
\description{
Markdown Task List
}
\examples{
md_task(c("Wake up", "Eat Breakfast", "Brush Teeth"), check = 1:2)
}
